crud-cloud-api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
slowapi==0.1.8
sqlalchemy==2.0.30
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pydantic==1.10.15
pytest==8.2.1
